Add another required long long cast to PRIM(time) 	Caused segfault

+++ prim-sys.cxx
@@ -310,7 +310,7 @@ PRIM(time) {
 
 	eprint(
 		"%6ldr %5ld.%ldu %5ld.%lds\t%L\n",
-		t1 - t0,
+		(long long) t1 - t0,
 		(long long) r.ru_utime.tv_sec, (long long) (r.ru_utime.tv_usec / 100000),
 		(long long) r.ru_stime.tv_sec, (long long) (r.ru_stime.tv_usec / 100000),
 		list.uget(), " "
